+++ app/controllers/companies_controller.rb
@@ -6,6 +6,7 @@ class CompaniesController < ApplicationController
 
   # GET /companies
   def index
+    authorize companies
     paginated_companies = companies.page(params[:page] || 1)
 
     render inertia: "Companies/Index", props: {
@@ -19,6 +20,7 @@ class CompaniesController < ApplicationController
 
   # GET /companies/:slug
   def show
+    authorize company
     if company.nil?
       render inertia: "Error", props: { status: 404 }
     else
@@ -30,6 +32,7 @@ class CompaniesController < ApplicationController
 
   # GET /companies/new
   def new
+    authorize Company
     render inertia: "Companies/New", props: {
       company: Company.new.render(view: :new)
     }
@@ -37,6 +40,7 @@ class CompaniesController < ApplicationController
 
   # GET /companies/:slug/edit
   def edit
+    authorize company
     render inertia: "Companies/Edit", props: {
       company: -> { company.render(view: :edit) }
     }
@@ -44,6 +48,7 @@ class CompaniesController < ApplicationController
 
   # POST /companies
   def create
+    authorize Company
     if Company.new(company_params).save
       # Assign admin permissions to user creating the record
       current_user.add_role :admin, company
@@ -57,6 +62,7 @@ class CompaniesController < ApplicationController
 
   # PATCH/PUT /companies/:slug
   def update
+    authorize company
     if company.update(company_params)
       redirect_to company, notice: 'Company was successfully updated.'
     else
@@ -66,6 +72,7 @@ class CompaniesController < ApplicationController
 
   # DELETE /companies/:slug
   def destroy
+    authorize company
     company.destroy
     respond_to do |format|
       format.html { redirect_to companies_url, notice: 'Company was successfully destroyed.' }
