+++ app/controllers/components_controller.rb
@@ -7,6 +7,7 @@ class ComponentsController < ApplicationController
 
   # GET /components
   def index
+    authorize components
     paginated_components = components.page(params[:page] || 1)
 
     render inertia: "Components/Index", props: {
@@ -29,6 +30,7 @@ class ComponentsController < ApplicationController
 
   # GET /components/:id
   def show
+    authorize component
     render inertia: "Components/Show", props: {
       component: -> { component.render(view: :show) }
     }
@@ -36,6 +38,7 @@ class ComponentsController < ApplicationController
 
   # GET /components/new
   def new
+    authorize Component
     render inertia: "Components/New", props: {
       component: Component.new.render(view: :new),
       models: -> { @active_company.models.find_by_category(:Component).render(view: :as_options) },
@@ -46,6 +49,7 @@ class ComponentsController < ApplicationController
 
   # GET /components/:id/edit
   def edit
+    authorize component
     render inertia: "Components/Edit", props: {
       component: component.render(view: :edit),
       models: -> { @active_company.models.find_by_category(:Component).render(view: :as_options) },
@@ -56,6 +60,7 @@ class ComponentsController < ApplicationController
 
   # GET /components/:id/checkout
   def checkout
+    authorize component
     redirect_to component if component.qty == 0
 
     assignment = Assignment.new
@@ -72,6 +77,7 @@ class ComponentsController < ApplicationController
 
   # GET /components/:id/checkin
   def checkin
+    authorize component
     redirect_to component if component.assignments.empty?
 
     assignment = component.assignment
@@ -88,6 +94,7 @@ class ComponentsController < ApplicationController
 
   # POST /components
   def create
+    authorize Component
     component.company = @active_company
     if component.save
       redirect_to component
@@ -98,6 +105,7 @@ class ComponentsController < ApplicationController
 
   # PATCH/PUT /components/:id
   def update
+    authorize component
     if component.update(component_params)
       redirect_to component
     else
@@ -107,6 +115,7 @@ class ComponentsController < ApplicationController
 
   # DELETE /components/:id
   def destroy
+    authorize component
     component.destroy
     respond_to do |format|
       format.html { redirect_to components_url, notice: "Component was successfully destroyed." }
