+++ app/controllers/networks_controller.rb
@@ -7,6 +7,7 @@ class NetworksController < ApplicationController
 
   # GET /networks
   def index
+    authorize networks
     paginated_networks = networks.page(params[:page] || 1)
 
     render inertia: "Networks/Index", props: {
@@ -20,6 +21,7 @@ class NetworksController < ApplicationController
 
   # GET /networks/:id
   def show
+    authorize network
     ips = IpLease.includes(:item).in_network(self.network)
 
     render inertia: "Networks/Show", props: {
@@ -34,6 +36,7 @@ class NetworksController < ApplicationController
 
   # GET /networks/new
   def new
+    authorize Network
     render inertia: "Networks/New", props: {
       network: -> { Network.new.render(view: :new) },
     }
@@ -41,6 +44,7 @@ class NetworksController < ApplicationController
 
   # GET /networks/:id/edit
   def edit
+    authorize network
     render inertia: "Networks/Edit", props: {
       network: -> { network.render(view: :edit) },
     }
@@ -48,6 +52,7 @@ class NetworksController < ApplicationController
 
   # POST /networks
   def create
+    authorize Network
     network.company = @active_company
     if network.save
       redirect_to network, notice: 'Network was successfully created'
@@ -58,6 +63,7 @@ class NetworksController < ApplicationController
 
   # PATCH/PUT /networks/:id
   def update
+    authorize network
     if network.update(network_params)
       redirect_to network, notice: 'Network was successfully updated'
     else
@@ -67,6 +73,7 @@ class NetworksController < ApplicationController
 
   # DELETE /networks/:id
   def destroy
+    authorize network
     network.destroy
     respond_to do |format|
       format.html { redirect_to networks_url, notice: 'Network was successfully destroyed.' }
