+++ app/controllers/people_controller.rb
@@ -8,6 +8,7 @@ class PeopleController < ApplicationController
 
   # GET /people
   def index
+    authorize people
     paginated_people = people.page(params[:page] || 1)
 
     render inertia: "People/Index", props: {
@@ -21,6 +22,7 @@ class PeopleController < ApplicationController
 
   # GET /people/1
   def show
+    authorize person
     render inertia: "People/Show", props: {
       person: person.render(view: :associations)
     }
@@ -28,6 +30,7 @@ class PeopleController < ApplicationController
 
   # GET /people/new
   def new
+    authorize Person
     render inertia: "People/New", props: {
       person: Person::AsCreate.new.render(view: :new),
       departments: InertiaRails.lazy(-> { @active_company.departments.render(view: :as_options) }),
@@ -38,6 +41,7 @@ class PeopleController < ApplicationController
 
   # GET /people/1/edit
   def edit
+    authorize person
     render inertia: "People/Edit", props: {
       person: person.render(view: :edit),
       departments: -> { @active_company.departments.render(view: :as_options) },
@@ -48,6 +52,7 @@ class PeopleController < ApplicationController
 
   # POST /people
   def create
+    authorize Person
     person = Person.new(handle_department_params)
 
     person.company = @active_company
@@ -61,6 +66,7 @@ class PeopleController < ApplicationController
 
   # PATCH/PUT /people/1
   def update
+    authorize person
     if person.update(handle_department_params)
       redirect_to person, notice: 'Person was successfully updated'
     else
@@ -70,6 +76,7 @@ class PeopleController < ApplicationController
 
   # DELETE /people/1
   def destroy
+    authorize person
     person.destroy
     redirect_to people_url, notice: 'Person was successfully destroyed.'
   end
